requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
python-dotenv==1.0.1
bcrypt==4.1.2
openai==1.14.3
